Patched Libstdc++, Linux-API-Headers and Ncurses to support Pass option

+++ Ncurses/6.2.0/6.2.0.cr
@@ -1,44 +1,78 @@
 class Target < ISM::Software
     
+    def prepare
+        if option("Pass1")
+            @buildDirectory = true
+            super
+
+            configureSource([] of String,buildDirectoryPath)
+            makeSource([Ism.settings.makeOptions,"-C","include"],buildDirectoryPath)
+            makeSource([Ism.settings.makeOptions,"-C","progs","tic"],buildDirectoryPath)
+        else
+            super
+        end
+    end
+
     def configure
         super
-        configureSource([   "--prefix=/usr",
-                            "--mandir=/usr/share/man",
-                            "--with-shared",
-                            "--without-debug",
-                            "--without-normal",
-                            "--enable-pc-files",
-                            "--enable-widec"],
-                            buildDirectoryPath)
+
+        if option("Pass1")
+            configureSource([   "--prefix=#{Ism.settings.rootPath}usr",
+                                "--host=#{Ism.settings.target}",
+                                "--build=$(./config.guess)",
+                                "--mandir=#{Ism.settings.rootPath}usr/share/man",
+                                "--with-manpage-format=normal",
+                                "--with-shared",
+                                "--without-debug",
+                                "--without-ada",
+                                "--without-normal",
+                                "--enable-widec"],
+                                buildDirectoryPath)
+        else
+            configureSource([   "--prefix=/usr",
+                                "--mandir=/usr/share/man",
+                                "--with-shared",
+                                "--without-debug",
+                                "--without-normal",
+                                "--enable-pc-files",
+                                "--enable-widec"],
+                                buildDirectoryPath)
+        end
     end
     
     def build
         super
+
         makeSource([Ism.settings.makeOptions],buildDirectoryPath)
     end
     
     def prepareInstallation
         super
-        makeSource([Ism.settings.makeOptions,"DESTDIR=#{builtSoftwareDirectoryPath}/#{Ism.settings.rootPath}","install"],buildDirectoryPath)
-        fileAppendData("#{builtSoftwareDirectoryPath(false)}/#{Ism.settings.rootPath}/usr/lib/libncurses.so","INPUT(-lncursesw)")
-        fileAppendData("#{builtSoftwareDirectoryPath(false)}/#{Ism.settings.rootPath}/usr/lib/libform.so","INPUT(-lformw)")
-        fileAppendData("#{builtSoftwareDirectoryPath(false)}/#{Ism.settings.rootPath}/usr/lib/libpanel.so","INPUT(-lpanelw)")
-        fileAppendData("#{builtSoftwareDirectoryPath(false)}/#{Ism.settings.rootPath}/usr/lib/libmenu.so","INPUT(-lmenuw)")
-        fileAppendData("#{builtSoftwareDirectoryPath(false)}/#{Ism.settings.rootPath}/usr/lib/libcursesw.so","INPUT(-lncursesw)")
-    end
 
-    def install
-        super
-        makeLink("ncursesw.pc","#{Ism.settings.rootPath}/usr/lib/pkgconfig/ncurses.pc",:symbolicLinkByOverwrite)
-        makeLink("formw.pc","#{Ism.settings.rootPath}/usr/lib/pkgconfig/form.pc",:symbolicLinkByOverwrite)
-        makeLink("panelw.pc","#{Ism.settings.rootPath}/usr/lib/pkgconfig/panel.pc",:symbolicLinkByOverwrite)
-        makeLink("menuw.pc","#{Ism.settings.rootPath}/usr/lib/pkgconfig/menu.pc",:symbolicLinkByOverwrite)
-        makeLink("libncurses.so","#{Ism.settings.rootPath}/usr/lib/libcurses.so",:symbolicLinkByOverwrite)
+        if option("Pass1")
+            makeSource([Ism.settings.makeOptions,"DESTDIR=#{builtSoftwareDirectoryPath}","TIC_PATH=$(pwd)/build/progs/tic","install"],buildDirectoryPath)
+            fileAppendData("#{builtSoftwareDirectoryPath}#{Ism.settings.rootPath}usr/lib/libncurses.so","INPUT(-lncursesw)")
+        else
+            makeSource([Ism.settings.makeOptions,"DESTDIR=#{builtSoftwareDirectoryPath}#{Ism.settings.rootPath}","install"],buildDirectoryPath)
+            fileAppendData("#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/libncurses.so","INPUT(-lncursesw)")
+            fileAppendData("#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/libform.so","INPUT(-lformw)")
+            fileAppendData("#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/libpanel.so","INPUT(-lpanelw)")
+            fileAppendData("#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/libmenu.so","INPUT(-lmenuw)")
+            fileAppendData("#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/libcursesw.so","INPUT(-lncursesw)")
+            makeLink("ncursesw.pc","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/pkgconfig/ncurses.pc",:symbolicLinkByOverwrite)
+            makeLink("formw.pc","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/pkgconfig/form.pc",:symbolicLinkByOverwrite)
+            makeLink("panelw.pc","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/pkgconfig/panel.pc",:symbolicLinkByOverwrite)
+            makeLink("menuw.pc","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/pkgconfig/menu.pc",:symbolicLinkByOverwrite)
+            makeLink("libncurses.so","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}usr/lib/libcurses.so",:symbolicLinkByOverwrite)
+        end
     end
 
     def clean
         super
-        deleteFile("#{Ism.settings.rootPath}/usr/lib/libncurses++w.a")
+
+        if !option("Pass1")
+            deleteFile("#{Ism.settings.rootPath}/usr/lib/libncurses++w.a")
+        end
     end
 
 end
